# docs/api_specification.yaml
...
openapi: 3.0.3
info:
  title: PumpIQ API
  description: AI-powered cryptocurrency trading recommendation system API
  version: 1.0.0
  contact:
    name: PumpIQ Support
    email: support@pumpiq.com
servers:
  - url: https://api.pumpiq.com/v1
    description: Production server
  - url: https://staging-api.pumpiq.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /recommendations:
    get:
      tags: [Recommendations]
      summary: Get recommendations for the user
      description: Get personalized AI-generated trading recommendations based on user preferences
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, STOPPED, EXPIRED]
            default: ACTIVE
        - name: recommendation_type
          in: query
          schema:
            type: string
            enum: [BUY, HOLD, SELL, AVOID]
        - name: min_confidence
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
        - name: risk_rating
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, VERY_HIGH]
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [SCALP, DAY_TRADE, SWING, LONG_TERM]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
    
    post:
      tags: [Recommendations]
      summary: Generate new recommendation
      description: Request AI to generate a new recommendation for a specific token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tokenId
              properties:
                tokenId:
                  type: string
                  format: uuid
                forceRefresh:
                  type: boolean
                  default: false
                  description: Force new analysis even if recent recommendation exists
      responses:
        '201':
          description: Recommendation generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
        '429':
          description: Rate limit exceeded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Token:
      type: object
      properties:
        tokenId:
          type: string
          format: uuid
        name:
          type: string
        ticker:
          type: string
        contractAddress:
          type: string
        blockchain:
          type: string
        currentPrice:
          type: number
        marketCap:
          type: number
        liquidity:
          type: number
        volume24h:
          type: number
        bondingCurvePercent:
          type: number
        isGraduated:
          type: boolean
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Recommendation:
      type: object
      properties:
        recommendationId:
          type: string
          format: uuid
        tokenId:
          type: string
          format: uuid
        token:
          $ref: '#/components/schemas/Token'
        recommendationType:
          type: string
          enum: [BUY, HOLD, SELL, AVOID]
        entryPrice:
          type: number
        currentPrice:
          type: number
        targets:
          type: object
          properties:
            target1:
              type: number
            target2:
              type: number
            target3:
              type: number
            stopLoss:
              type: number
        confidenceScore:
          type: number
        riskRating:
          type: string
        timeframe:
          type: string
        dataContribution:
          type: object
          properties:
            news:
              type: boolean
            onchain:
              type: boolean
            technical:
              type: boolean
            social:
              type: boolean
        aiReasoning:
          type: string
        keyFactors:
          type: array
          items:
            type: object
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...